Develop

+++ app/Http/Controllers/Api/V1/BillingController.php
@@ -7,24 +7,54 @@ use Illuminate\Http\Request;
 use App\Models\Billing;
 use App\Http\Resources\Api\V1\BillingResource;
 use Validator;
+use Carbon\Carbon;
 
 class BillingController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
+        $search = $request->search ?? null;
+        $start_date = $request->start_date ?? Carbon::now()->startOfMonth();
+        $end_date = $request->end_date ?? Carbon::now();
+
         $user = auth()->user();
-        // $billings = Billing::where('user_id', $user->id)->get();
-        // list billing by user_id and destination ordered from visit, promise, pay
-        $billings = Billing::where('destination', 'visit')
+
+        if ($request->start_date && $request->end_date) {
+            $start_date = Carbon::parse($request->start_date)->format('Y-m-d');
+            $end_date = Carbon::parse($request->end_date)->format('Y-m-d');
+        }
+
+        $billings = Billing::whereBetween('date', [$start_date, $end_date])
+            ->where('user_id', $user->id)
+            ->where('destination', 'visit')
             ->orWhere('destination', 'promise')
             ->orWhere('destination', 'pay')
-            ->where('user_id', $user->id)
             ->orderBy('destination', 'asc')
             ->orderBy('date', 'asc')
             ->get();
+
+        if ($search) {
+            $billings = Billing::with('customer', 'user')
+            ->whereBetween('date', [$start_date, $end_date])
+            ->where('user_id', $user->id)
+            ->where('destination', 'like', '%' . $search . '%')
+            ->orWhere('no_billing', 'like', '%' . $search . '%')
+            ->orWhereHas('customer', function($q) use($search, $user, $start_date, $end_date) {
+                $q->where('name_customer', 'like', '%' . $search . '%')
+                    ->orWhere('no', 'like', '%' . $search . '%')
+                    ->whereHas('billing', function($q) use($user, $start_date, $end_date) {
+                        $q->whereBetween('date', [$start_date, $end_date])
+                            ->where('user_id', $user->id);
+                    });
+            })
+            ->orWhere('destination', 'like', '%' . $search . '%')
+            ->orderBy('destination', 'asc')
+            ->orderBy('date', 'asc')
+            ->get();
+        }
 
         return response()->json([
             'status' => 'success',
@@ -84,7 +114,8 @@ class BillingController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $billing = Billing::find($id);
+        $user = auth()->user();
+        $billing = Billing::where('user_id', $user->id)->where('id', $id)->first();
 
         if (!$billing) {
             return response()->json([
@@ -96,15 +127,24 @@ class BillingController extends Controller
         $validator = Validator::make($request->all(), [
             'no_billing' => 'required|unique:billings,no_billing,' . $id,
             'date' => 'required|date',
-            'bank_account_id' => 'required|exists:bank_accounts,id',
-            'user_id' => 'required|exists:users,id',
+            // 'bank_account_id' => 'required|exists:bank_accounts,id',
+            // 'user_id' => 'required|exists:users,id',
             'destination' => 'required|in:visit,promise,pay',
-            'result' => 'nullable',
-            'promise_date' => 'required_if:destination,promise',
-            'amount' => 'required_if:destination,pay',
-            'image_amount' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-            'signature_officer' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-            'signature_customer' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            // 'image_visit' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            'image_visit' => 'nullable',
+            'description_visit' => 'nullable',
+            'promise_date' => 'nullable',
+            // 'image_promise' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            'image_promise' => 'nullable',
+            'description_promise' => 'nullable',
+            'amount' => 'nullable',
+            // 'image_amount' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            'image_amount' => 'nullable',
+            'description_amount' => 'nullable',
+            // 'signature_officer' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            // 'signature_customer' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            'signature_officer' => 'nullable',
+            'signature_customer' => 'nullable',
         ]);
 
         if ($validator->fails()) {
@@ -116,11 +156,40 @@ class BillingController extends Controller
         }
 
         $validatedData = $validator->validated();
+        $validatedData['updated_by'] = $user->id;
+
+        // save image to public/images/billings and change name to timestamp
+        if ($request->hasFile('image_visit')) {
+            // remove old image
+            if ($billing->image_visit != null && file_exists(public_path('images/billings/' . $billing->image_visit))) {
+                unlink(public_path('images/billings/' . $billing->image_visit));
+            }
+
+            // save image to public/images/billings and change name file to name user-timestamp
+            $file = $request->file('image_visit');
+            $fileName = $validatedData['no_billing'] . '-' . 'visit' . '-' . time() . '.' . $file->getClientOriginalExtension();
+            $file->move(public_path('images/billings'), $fileName);
+            $validatedData['image_visit'] = $fileName;
+        }
+
+        // save image to public/images/billings and change name to timestamp
+        if ($request->hasFile('image_promise')) {
+            // remove old image
+            if ($billing->image_promise != null && file_exists(public_path('images/billings/' . $billing->image_promise))) {
+                unlink(public_path('images/billings/' . $billing->image_promise));
+            }
+
+            // save image to public/images/billings and change name file to name user-timestamp
+            $file = $request->file('image_promise');
+            $fileName = $validatedData['no_billing'] . '-' . 'promise' . '-' . time() . '.' . $file->getClientOriginalExtension();
+            $file->move(public_path('images/billings'), $fileName);
+            $validatedData['image_promise'] = $fileName;
+        }
 
         // save image to public/images/billings and change name to timestamp
         if ($request->hasFile('image_amount')) {
             // remove old image
-            if (file_exists(public_path('images/billings/' . $billing->image_amount))) {
+            if ($billing->image_amount != null && file_exists(public_path('images/billings/' . $billing->image_amount))) {
                 unlink(public_path('images/billings/' . $billing->image_amount));
             }
 
@@ -134,7 +203,7 @@ class BillingController extends Controller
         // save image to public/images/billings and change name to timestamp
         if ($request->hasFile('signature_officer')) {
             // remove old image
-            if (file_exists(public_path('images/billings/' . $billing->signature_officer))) {
+            if ($billing->signature_officer != null && file_exists(public_path('images/billings/' . $billing->signature_officer))) {
                 unlink(public_path('images/billings/' . $billing->signature_officer));
             }
 
@@ -148,7 +217,7 @@ class BillingController extends Controller
         // save image to public/images/billings and change name to timestamp
         if ($request->hasFile('signature_customer')) {
             // remove old image
-            if (file_exists(public_path('images/billings/' . $billing->signature_customer))) {
+            if ($billing->signature_customer != null && file_exists(public_path('images/billings/' . $billing->signature_customer))) {
                 unlink(public_path('images/billings/' . $billing->signature_customer));
             }
 
@@ -160,23 +229,23 @@ class BillingController extends Controller
         }
 
         $validatedData['updated_by'] = auth()->id();
-        if ($request->destination == 'visit') {
-            $validatedData['result'] = null;
-            $validatedData['promise_date'] = null;
-            $validatedData['amount'] = null;
-            if ($billing->image_amount != null && file_exists(public_path('images/billings/' . $billing->image_amount))) {
-                unlink(public_path('images/billings/' . $billing->image_amount));
-            }
-            $validatedData['image_amount'] = null;
-            if ($billing->siganture_officer != null && file_exists(public_path('images/billings/' . $billing->signature_officer))) {
-                unlink(public_path('images/billings/' . $billing->signature_officer));
-            }
-            $validatedData['signature_officer'] = null;
-            if ($billing->siganture_customer != null && file_exists(public_path('images/billings/' . $billing->signature_customer))) {
-                unlink(public_path('images/billings/' . $billing->signature_customer));
-            }
-            $validatedData['signature_customer'] = null;
-        }
+        // if ($request->destination == 'visit') {
+        //     $validatedData['description_visit'] = null;
+        //     $validatedData['promise_date'] = null;
+        //     $validatedData['amount'] = null;
+        //     if ($billing->image_amount != null && file_exists(public_path('images/billings/' . $billing->image_amount))) {
+        //         unlink(public_path('images/billings/' . $billing->image_amount));
+        //     }
+        //     $validatedData['image_amount'] = null;
+        //     if ($billing->siganture_officer != null && file_exists(public_path('images/billings/' . $billing->signature_officer))) {
+        //         unlink(public_path('images/billings/' . $billing->signature_officer));
+        //     }
+        //     $validatedData['signature_officer'] = null;
+        //     if ($billing->siganture_customer != null && file_exists(public_path('images/billings/' . $billing->signature_customer))) {
+        //         unlink(public_path('images/billings/' . $billing->signature_customer));
+        //     }
+        //     $validatedData['signature_customer'] = null;
+        // }
 
         $billing->update($validatedData);
